reorganize program structure
