--- programs/seasonal_dec/Cargo.toml
@@ -1,21 +1,0 @@
-[package]
-name = "seasonal_dec"
-version = "0.1.0"
-description = "Created with Anchor"
-edition = "2021"
-
-[lib]
-crate-type = ["cdylib", "lib"]
-name = "seasonal_dec"
-
-[features]
-default = []
-cpi = ["no-entrypoint"]
-no-entrypoint = []
-no-idl = []
-no-log-ix-name = []
-idl-build = ["anchor-lang/idl-build", "anchor-spl/idl-build"]
-
-[dependencies]
-anchor-lang = "0.30.1"
-anchor-spl = "0.30.1"
